move the constructor inside
