+++ TileMap.ts
@@ -1,9 +1,38 @@
-import { TileInterface } from './Tile';
+import * as pc from 'playcanvas';
+import { Tile } from './Tile';
 
 export class TileMap {
-    private tiles: { [key: string]: TileInterface } = {};
+    private tiles: { [key: string]: Tile } = {};
+    private center: pc.Vec3;
 
-    registerTile(key: string, tile: TileInterface): void {
+    constructor(app: pc.Application) {
+        this.initialize(app).then(() => {
+            console.log("TileMap initialized with tiles.");
+        }).catch(err => {
+            console.error("Failed to initialize TileMap:", err);
+        });
+    }
+
+    async initialize(app: pc.Application) {
+        const colorResponse = await fetch('./colors.json');
+        const positionResponse = await fetch('./positions.json');
+        const colorsJson = await colorResponse.json();
+        const positionsJson = await positionResponse.json();
+
+        const initColors: { [key: string]: pc.Color } = {};
+        Object.keys(colorsJson).forEach(key => {
+            const [r, g, b] = colorsJson[key];
+            initColors[key] = new pc.Color(r, g, b);
+        });
+        Object.keys(positionsJson).forEach(key => {
+            const [x, z] = positionsJson[key];
+            const posVec = new pc.Vec3(x, -0.5, z);
+            // this.registerTile(key, new Tile(key, this, app, posVec, initColors[key]));
+            new Tile(key, this, app, posVec, initColors[key]);
+        });
+    }
+
+    registerTile(key: string, tile: Tile): void {
         if (this.tiles[key]) {
             this.tiles[key].remove();  // Ensure no duplicates
         }
@@ -22,11 +51,11 @@ export class TileMap {
         }
     }
 
-    getTile(key: string): TileInterface {
+    getTile(key: string): Tile {
         return this.tiles[key];
     }
 
-    getTiles(): { [key: string]: TileInterface } | undefined {
+    getTiles(): { [key: string]: Tile } | undefined {
         return this.tiles;
     }
 }
